+++ changelog.md
@@ -1,5 +1,14 @@
 # Application Framework Changelog
 
+## v5.9.0 - Country management
+- Wizards: Added the possibility to specify step classes, loosening the class structure.
+- Countries: Added the country management screens.
+- Countries: Added user rights to manage countries.
+- Countries: Improved country ISO code alias handling (`uk` and `gb`).
+- Dependencies: Updated AppUtils Core to [v2.3.11](https://github.com/Mistralys/application-utils-core/releases/tag/2.3.11).
+- Dependencies: Updated AppUtils Collections to [v1.1.5](https://github.com/Mistralys/application-utils-collections/releases/tag/1.1.5).
+- Dependencies: Updated AppLocalize to [v2.0.0](https://github.com/Mistralys/application-localization/releases/tag/2.0.0).
+
 ## v5.8.3 - Fixes and PHP8.4 
 - Session: Fixed the CLI session handling to avoid errors when running in CLI mode.
 - PHP8.4: Fixed deprecation warnings for implicit nullable method arguments.